L2 Cache 优化。Copysign
Case1: Pass, Result: 5.794
Case2: Pass, Result: 24.4696
Case3: Pass, Result: 8.0784
Case4: Pass, Result: 27.1012
Case5: Pass, Result: 811.4752
prof_sum: 876.9184

+++ Copysign/op_host/copysign_tiling.h
@@ -3,18 +3,18 @@
 
 namespace optiling {
 BEGIN_TILING_DATA_DEF(CopysignTilingData)
-TILING_DATA_FIELD_DEF(int64_t, former_length);
-TILING_DATA_FIELD_DEF(int64_t, tail_length);
-TILING_DATA_FIELD_DEF(int64_t, former_num);
-TILING_DATA_FIELD_DEF(int64_t, tile_length);
+TILING_DATA_FIELD_DEF(int32_t, former_length);
+TILING_DATA_FIELD_DEF(int32_t, tail_length);
+TILING_DATA_FIELD_DEF(int32_t, former_num);
+TILING_DATA_FIELD_DEF(int32_t, tile_length);
 
-TILING_DATA_FIELD_DEF(int64_t, vector_length);
-TILING_DATA_FIELD_DEF(int64_t, align_vector_length);
+TILING_DATA_FIELD_DEF(int32_t, vector_length);
+TILING_DATA_FIELD_DEF(int32_t, align_vector_length);
 
 TILING_DATA_FIELD_DEF(uint8_t, board_cast);
-TILING_DATA_FIELD_DEF(int64_t, out_dim1);
-TILING_DATA_FIELD_DEF(int64_t, out_dim2);
-TILING_DATA_FIELD_DEF(int64_t, out_dim3);
+TILING_DATA_FIELD_DEF(int32_t, out_dim1);
+TILING_DATA_FIELD_DEF(int32_t, out_dim2);
+TILING_DATA_FIELD_DEF(int32_t, out_dim3);
 END_TILING_DATA_DEF;
 
 REGISTER_TILING_DATA_CLASS(Copysign, CopysignTilingData)
